requirements.txt
asgiref==3.8.1
Django==5.1.3
django-ckeditor==6.7.1
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
django-js-asset==2.2.0
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.1
